Removed all input validations from middleware, all schemas are now validated through their respective api routes using utility function

+++ src/app/api/login/route.ts
@@ -3,14 +3,27 @@ import * as bcrypt from 'bcrypt';
 import * as jose from 'jose';
 import { NextResponse } from 'next/server';
 import { cookies } from 'next/headers'
-import {env} from '../../../utility/EnvironmentValidatior';
+import { env } from '../../../utility/EnvironmentValidatior';
 import { User } from '@/types/TableModels';
+import { object, string } from 'yup';
+import ValidationUtility from '@/utility/ValidateorUtility';
 export async function POST(request: Request) {
-    console.log("bypassed");
+    const data = await request.json();
+    const schema = object({
+        username: string().required().min(1),
+        password: string().required().min(1),
+    });
+    const validationResult = await ValidationUtility(schema, data);
+    if (validationResult.status === "fail") {
+        return NextResponse.json(validationResult);
+    }
+
+
+
+
     const database = getConnection();
     if (database) {
-        let info = await request.json();
-        let data = info.schema;
+
         const { username, password } = data;
         const user = await database<User>("users").where({ username: username }).first();
         await database.destroy();
@@ -23,7 +36,7 @@ export async function POST(request: Request) {
         if (await bcrypt.compare(password, user.password)) {
 
             const secret = new TextEncoder().encode(
-                env.SECRET_KEY
+                env.USER_SECRET_KEY
             );
             const token = await new jose.SignJWT({ username: username, password: password })
                 .setProtectedHeader({ alg: "HS256" })
